requirements.txt
moderngl==5.9.0
numpy==1.25.2
pygame_ce==2.3.1
